feat(segcache): ABA-safe 64-bit CAS tokens
Replace the u32 location-truncated CAS token with a 64-bit token
combining the 44-bit packed location with the segment's 16-bit
generation counter (bits 59..44), ported from crucible's CasToken.

The old token had a false-success case: (a) truncating the 44-bit
location to u32 aliases locations whose seg_ids differ by a multiple
of 4096, and (b) even at full width, a recycled segment can place the
same key at the same offset, letting a stale token match a value the
client never observed. The generation counter, bumped on every
segment recycle, distinguishes lifetimes of the same location.

The 64-bit width also matches the memcache protocol's "cas unique".

+++ crates/segcache/src/item/mod.rs
@@ -9,12 +9,12 @@ pub(crate) use reserved::ReservedItem;
 
 /// The base unit of data returned by a cache lookup.
 pub struct Item {
-    cas: u32,
+    cas: u64,
     raw: RawItem,
 }
 
 impl Item {
-    pub(crate) fn new(raw: RawItem, cas: u32) -> Self {
+    pub(crate) fn new(raw: RawItem, cas: u64) -> Self {
         Item { cas, raw }
     }
 
@@ -33,8 +33,11 @@ impl Item {
         self.raw.value()
     }
 
-    /// CAS value for the item
-    pub fn cas(&self) -> u32 {
+    /// CAS value for the item, matching the memcache protocol's 64-bit
+    /// "cas unique". Combines the item's location with its segment's
+    /// generation counter; any update, relocation, or segment recycle
+    /// invalidates outstanding values.
+    pub fn cas(&self) -> u64 {
         self.cas
     }
